ai-media-system/requirements.txt
beautifulsoup4==4.12.2
playwright==1.40.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai-whisper==20231117
yt-dlp==2023.11.16
